mutt: Initial version of mail2html.py script.
Implemented script send_mail.py for sending emails by using msmtp.
Before sending email we will try to convert it to html if marker `!m` is
used in the beginning of the message or in the end.

Implemented preview_mail.py. This script helps to show preview of html
email.